Pike: Update Macmini data to stop EFI 1.6 updates
String number of the BIOS release date. The date string, if supplied, is in either mm/dd/yy or mm/dd/yyyy format. If the year portion of the string is two digits, the year is assumed to be 19yy.

NOTE: The mm/dd/yyyy format is required for SMBIOS version 2.3 and later.